internal/libhive: move types to shared package
